requirements.txt
readchar~=4.0.5
pyserial~=3.5
numpy~=1.24.4
PyYAML~=6.0.1
imgui-bundle==1.6.3